TFLite/requirement.txt
pycocotools
tensorflow
numpy
opendatasets
kaggle
pillow